Linux RH devtoolset for gnu-7, changes LD_LIBRARY_PATH and PATH
This brings in the new GNU compiler. If the standard GNU compiler
on Linux is too old to build WRF, this is what is required. The
side effect is that the "yum install netcdf" commands are now
generating inconsistent Fortran *.mod files. The "yum install"
command is defaulting to an assumed GNU version (4.8.5, OLD).
With some source changes in WRF, there is no longer a need for
the Fortran MODULE "use netcdf" statment, we can get away with
"include netcdf.inc".