Add intermittent print so travis doesn't think process is hung, and stop the build
This is due to the travis requirement that the std out and err output from the
configure and make of the external libs (netcdf, hdf5, openmpi) was too large.
So, you stick the std out and err in a file to be a nice guy. Is travis happy?
No. Now it sees no output for a long time (since the three libs take a while) and says
"Hey, I'll bet this process is just hung".

Travis wants to see output every 10 minutes. Instead of modifying the .travis.yml
file, I tried to manufacture output after each of the configures (4 of them) and
after each of the makes (4 of them). Fingers crossed.

Sheesh.

Changes to be committed:
	modified:   Dockerfile
	modified:   Dockerfile-template